Backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23

# AI & Embeddings - UPDATED VERSIONS
langchain==0.1.16
langchain-openai==0.1.1
langchain-community==0.0.29
openai==1.14.0

# Vector Database
chromadb==0.4.24

# Utilities
python-dotenv==1.0.0
pydantic==2.6.4